Fixes from Jenkins.

+++ lib/SimpleHttpClient/SimpleHttpClient.cpp
@@ -603,7 +603,7 @@ void SimpleHttpClient::setRequest(rest::RequestType method, std::string const& l
 
   if (method != rest::RequestType::GET) {
     _writeBuffer.appendText(TRI_CHAR_LENGTH_PAIR("Content-Length: "));
-    _writeBuffer.appendInteger(bodyLength);
+    _writeBuffer.appendInteger(static_cast<uint64_t>(bodyLength));
     _writeBuffer.appendText(TRI_CHAR_LENGTH_PAIR("\r\n\r\n"));
   } else {
     _writeBuffer.appendText(TRI_CHAR_LENGTH_PAIR("\r\n"));
